chore: split jest.setup.ts into multiple files
