+++ web/jest.config.js
@@ -5,7 +5,7 @@ module.exports = {
   rootDir: 'src/',
   roots: ['<rootDir>', '<rootDir>/../server'],
   testPathIgnorePatterns: ['.next/', 'node_modules/'],
-  setupFilesAfterEnv: ['../jest.setup.defineProperty.ts', '../jest.setup.ts', './test/matchers/index.ts'],
+  setupFilesAfterEnv: ['../jest.setup.defineProperty.ts', '../jest.setup.ts'],
   transform: {
     '^.+\\.[t|j]sx?$': 'babel-jest',
   },
