+++ web/src/test/setup/mockCopy.ts
@@ -1,0 +1,6 @@
+jest.mock('copy-to-clipboard', () => {
+  (global as any).copyClipboardMock = jest.fn();
+  return (global as any).copyClipboardMock;
+});
+
+export {};
