+++ web/src/test/setup/mockFetch.ts
@@ -1,0 +1,6 @@
+// import `fetch` to fix apollo-link-http-common's Invariant Violation:
+// fetch is not found globally and no fetcher passed, to fix pass a fetch for
+// your environment like https://www.npmjs.com/package/node-fetch.
+(global as any).fetch = require('node-fetch');
+
+export {};
