+++ web/src/test/setup/mockGa.ts
@@ -1,0 +1,6 @@
+// Google analytics mock
+jest.mock('react-ga', () => {
+  return { initialize: jest.fn(), set: jest.fn(), pageview: jest.fn(), event: jest.fn() };
+});
+
+export {};
