+++ web/src/test/setup/mockPlay.ts
@@ -1,0 +1,7 @@
+// mock functions for HTMLMediaElement
+// https://github.com/jsdom/jsdom/issues/2155#issuecomment-366703395
+(window as any).HTMLMediaElement.prototype.play = () => {
+  /* do nothing */
+};
+
+export {};
