+++ web/src/test/setup/mockScrollTo.ts
@@ -1,0 +1,4 @@
+// mock window.scrollTo
+(global as any).scrollTo = jest.fn();
+
+export {};
