+++ web/src/test/setup/mockSw.ts
@@ -1,0 +1,12 @@
+if (typeof window !== 'undefined') {
+  let swRegisterMock = jest.fn();
+  let mockPromise = new Promise(function (resolve) {
+    resolve({ update: jest.fn() });
+  });
+  swRegisterMock.mockReturnValue(mockPromise);
+  (global as any).navigator.serviceWorker = {
+    register: swRegisterMock,
+  };
+}
+
+export {};
