+++ web/src/test/setup/mockWindowLocation.ts
@@ -1,0 +1,5 @@
+const { location } = global as any;
+delete (global as any).location;
+(global as any).location = { ...location, reload: jest.fn() };
+
+export {};
